Move tests to new folder to mirror recent repackaging of some classes
