+++ core/src/test/java/com/klaviyo/coresdk/networking/requests/NetworkRequestTest.kt
@@ -1,4 +1,4 @@
-package com.klaviyo.coresdk.networking
+package com.klaviyo.coresdk.networking.requests
 
 import android.content.Context
 import android.net.ConnectivityManager
@@ -7,7 +7,7 @@ import android.net.NetworkCapabilities
 import android.net.NetworkInfo
 import androidx.test.filters.SdkSuppress
 import com.klaviyo.coresdk.KlaviyoConfig
-import com.klaviyo.coresdk.networking.requests.NetworkRequest
+import com.klaviyo.coresdk.networking.RequestMethod
 import com.nhaarman.mockitokotlin2.*
 import org.junit.Assert.assertEquals
 import org.junit.Before
